skills/deako.mark2/requirements.txt
telnetlib3
zeroconf